backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai
gspread
oauth2client
python-multipart
requests
twilio
google-cloud-texttospeech
jinja2
chromadb
pysqlite3-binary
